+++ src/routes/api/auth/index.js
@@ -2,7 +2,7 @@
 
 const express = require('express')
 const router = express.Router()
-const authController = require('../../controllers/auth.controller')
+const authController = require('../../../controllers/auth.controller')
 
 /**
  * @api {post} /api/auth/login Login
